ai/shared/pyproject.toml
[project]
name = "civitro-common"
version = "0.1.0"
description = "Shared library for Civitro Python microservices"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "asyncpg>=0.30.0",
    "motor>=3.6.0",
    "aioredis>=2.0.0",
    "aiokafka>=0.11.0",
    "aioboto3>=13.0.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["civitro_common*"]
